update today mile and kcal

+++ wego/data/models.py
@@ -9,9 +9,9 @@ from adminset.models import Users
 class DayData(models.Model):
     user = models.ForeignKey(User, on_delete=models.SET_NULL, null=True)
     step = models.IntegerField(u"步数", default=0)
-    mileage = models.DecimalField(u"距离", default=0, max_digits=25, decimal_places=2, )
-    altitude = models.DecimalField(u"海拔", default=0, max_digits=25, decimal_places=2, )
-    calorie = models.DecimalField(u"卡路里", default=0, max_digits=25, decimal_places=2, )
+    mileage = models.FloatField(u"距离", default=0)
+    altitude = models.FloatField(u"海拔", default=0)
+    calorie = models.IntegerField(u"卡路里", default=0)
     created_time = models.DateField()
 
     def get_user_info(self):
